# src/main/resources/swagger/api.yml
swagger: '2.0'
info:
  title: Restaurant Service
  description: >-
    This is Swagger UI environment generated for the Restaurant Service
  version: '1.0'
host: serverRoot
basePath: /v1/
schemes:
  - https
paths:
...
  '/restaurants/search':
    get:
      parameters:
        - in: query
          name: lat
          type: number
          format: double
          description: your latitude position 
        - in: query
          name: long
          type: number
          format: double
          description: your longitude position 
        - in: query
          name: distance
          type: integer
          description: max distance from your position       
      produces:
        - application/json
        
      operationId: restaurantsSearch
      summary: Find  restaurantes near you
      tags:
        - restaurants-api-controller
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/RestaurantDTO'          
        '400':
          description: |-
            Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not allowed
        '409':
          description: Conflict
        '500':
          description: Internal Server Error
definitions:
  RestaurantDTO:
    type: object
    description: Restaurant Object
    required: 
      - name
      - description
      - lat
      - long
    properties:
      id:
        type: string
        format: uuid
        description: Unique number for object
      name:
        type: string
        description: Restaurant Name
      description:
        type: string
        description: Restaurant Description
      latitude:
        type: number
        format: double
        description: Latitude 
      longitude:
        type: number
        format: double
        description: Longitude 
      rating:
        type: integer
        minimum: 1
        maximum: 10
        description: Restaurant rating (1-10)
    example:
     id: 721d68a7-f6bb-4863-b071-0e58a50da5ae
     name: Da Lucia
     description: Restaurante italiano
     lat: 52.498890 
     long: 13.342076
     rating: 8